Adds protocol deletion functionality
Implements a feature to allow users to mark protocols as deleted, rather than permanently removing them from the database.

This change introduces a `hidden_user` flag to the `intra_edivi` table, indicating if a protocol has been marked as deleted by a user. The list view now displays a "G" badge for protocols marked as deleted. A notification is also shown if a protocol has been deleted.

It provides a button to delete all non released protocols.

Includes database migration to add the `hidden_user` column.