Enhances circulatory assessment in first report
Adds fields for 'recap time' and 'strong bleeding' to the circulatory assessment section of the first report.

This update improves the comprehensiveness of the circulatory assessment by
including additional key indicators, and moves EKG out of circulation, since
this is a function.

Additionally, applies minor UI adjustments, such as increasing the width of
validation indicator bars and adjusting the color of checked boxes.