Disables transport for specific patient categories
Prevents transport assignment for patients triaged as SK4, SK5, or SK6.

This change ensures that patients with these specific categories are not offered transport options in the system,
aligning with updated triage protocols.  The transport section is now conditionally rendered based on the patient's
category. Also adjustes the board view to prevent transport assignments for said categories.

Related to #199